separated code into 2 files
