+++ client/app/page.tsx
@@ -2,17 +2,15 @@
 
 import React, { useEffect, useState } from 'react';
 import { QRCodeSVG } from 'qrcode.react';
-import { BiCopy } from 'react-icons/bi';
 import './style.css'
 import * as crpt from './encryption';
 import * as pool from './pool';
+import Chat from './chat';
 
 
 const serverUrl = "ws://localhost:8080/ws";
 
 export default function Home() {
-  const [message, setMessage] = useState('');
-  const [copiedMessage, setCopiedMessage] = useState('');
   const [history, setHistory] = useState<string[]>([]); 
   const [status, setStatus] = useState('disconnected');
   const [socket, setSocket] = useState<WebSocket | null>(null);
@@ -26,7 +24,7 @@ export default function Home() {
   const [pairingAccepted, setPairingAccepted] = useState(false);
   const [peerAccepted, setPeerAccepted] = useState(false);
   const [clientName, setClientName] = useState('');
-  const [selectetdFile, setSelectedFile] = useState([]);
+  const [Disconnect, setDisconnect] = useState(false);
   var receivedChunks: { [key: number]: ArrayBuffer } = {};
   const workerPool = new pool.Pool(5);
 
@@ -57,24 +55,27 @@ export default function Home() {
       sendClientInfo();
   }, [peerPublicKey, socket]);
 
-
-  const disconnectClient = () => {
-    if (socket)
+  useEffect(() => {
+    if (Disconnect) {
+      if (socket)
       socket.close();
-    setHistory([]);
-    setStatus('disconnected');
-    setClientId(null);
-    setChatStarted(false);
-    setKeyPair(null);
-    setPeerPublicKey(null);
-    setIsConnecting(false);
-    setPairInfo("");
-    setPairingAccepted(false);
-    setPeerAccepted(false);
-    setSelectedFile([]);
-    receivedChunks = {};
-    setClientName('');
-  }
+      if (!chatStarted)
+      {
+        console.log("chat not started");
+        setHistory([]);
+      }
+      setStatus('disconnected');
+      setClientId(null);
+      setKeyPair(null);
+      setPeerPublicKey(null);
+      setIsConnecting(false);
+      setPairInfo("");
+      setPairingAccepted(false);
+      setPeerAccepted(false);
+      receivedChunks = {};
+      setClientName('');
+    }
+  }, [Disconnect]);
 
   const generateKeyPair = async () => {
     const keyPair = await crpt.generateKeyPair();
@@ -88,22 +89,6 @@ export default function Home() {
     
     return adjectives[Math.floor(Math.random() * adjectives.length)] + ' ' + nouns[Math.floor(Math.random() * nouns.length)];
   }
-
-
-  const onFileChange = (e: any) => {
-    const maxFileSize = 200 * 1024 * 1024;
-    if (e.target.files[0] && e.target.files[0].size > maxFileSize) {
-      alert('Max file size is ' + maxFileSize / 1024 / 1024 + 'MB.')
-      e.target.value = null;
-    }
-    else {
-      setSelectedFile(e.target.files);
-      console.log(e.target.files[0]);
-      console.log(e.target.files[0].name);
-      console.log(e.target.files[0].size);
-      console.log(e.target.files[0].type);
-    }
-  };
 
 
   const getUrl = () => {
@@ -122,7 +107,6 @@ export default function Home() {
 
     // Connection opened
     socket.addEventListener('open', async (event) => {
-      console.log('WebSocket connection established.');
       setError("");
     });
 
@@ -134,7 +118,6 @@ export default function Home() {
       } else if (event.data.slice(0, 13) === 'client-info: ') {
         let encrypted = event.data.slice(13);
         let decrypted = await crpt.decrypt(encrypted, keyPair);
-        console.log('Decrypted client info: ', decrypted);
         setPairInfo(decrypted);
 
       } else if (event.data === 'connected') {
@@ -142,11 +125,9 @@ export default function Home() {
 
       } else if (event.data.slice(0,11) === 'client-id: '){
         let clientId = event.data.slice(11);
-        console.log('Received client id: ', clientId);
         setClientId(clientId);
 
       } else if (event.data.slice(0, 12) === 'public-key: ') {
-        console.log('Received public key: ', event.data.slice(12));
         let key = event.data.slice(12);
         setPeerPublicKey(await crpt.importAndSetPeerPublicKey(key));
 
@@ -165,7 +146,7 @@ export default function Home() {
         receiveFileByChunks(encrypted);
       }
       else{
-        console.log('Received unknown message: ', event.data);
+        //console.log('Received unknown message: ', event.data);
       }
     });
 
@@ -174,7 +155,6 @@ export default function Home() {
     });    
 
     socket.addEventListener('close', (event) => {
-      console.log('WebSocket connection closed: ', event.reason);
       if (event.reason === 'Client with this ID does not exist')
       {
         setError("id-not-found");
@@ -183,7 +163,7 @@ export default function Home() {
       } else {
         setError("server-error");
       }
-      disconnectClient();
+      setDisconnect(true);
     });};
 
     const receiveFileByChunks = async (encrypted: string) => {
@@ -236,92 +216,10 @@ export default function Home() {
     }
 
     const rejectPairing = async () => {
-      disconnectClient();
+      setDisconnect(true);
       setError("Pairing was rejected. Try again.");
     }
 
-
-  const sendMessage = async (text: string) => {
-    if (!socket || socket.readyState !== WebSocket.OPEN) {
-      return;
-    }
-
-    if (!text.trim()) {
-      return;
-    }
-    
-    setHistory((prev: any) => [...prev, 'You: ' + text]);
-
-    let encryptedString = await crpt.encrypt(text, peerPublicKey);
-    socket.send("msg: " + encryptedString);
-    setMessage('');
-    setChatStarted(true);
-  };
-
-
-  const handleSendMessage = async () => {
-    await sendMessage(message);
-  };
-
-  const handleSendFile = async () => {
-    console.log(new Date().toLocaleTimeString())
-    const worker = new Worker(new URL('./workers/send-file.ts', import.meta.url));
-
-  // Define the event listener to handle messages from the worker
-  worker.onmessage = function (event) {
-    socket!.send(event.data);
-  };
-
-  // Send data to the worker
-  const data = {file: selectetdFile[0], peerPublicKey: peerPublicKey};
-  worker.postMessage(data);
-
-  // Cleanup the worker when the component unmounts
-  return () => {
-    worker.terminate();
-  };
-};
-
-
-  const clearMessages = () => {
-    setHistory([]);
-    setChatStarted(false);
-  };
-
-
-  const copyToClipboard = (message: string) => {
-    const messageWithoutPrefix = message.replace(/^(You|Peer):\s*/, '');
-    navigator.clipboard.writeText(messageWithoutPrefix);
-    setCopiedMessage(message);
-  };
-
-
-  const handlePaste = async () => {
-    try {
-      const permissionStatus = await navigator.permissions.query({name: 'clipboard-read' as PermissionName});
-      switch (permissionStatus.state) {
-        case "granted":
-          const text = await navigator.clipboard.readText();
-          sendMessage(text);
-          break;
-        case "prompt":
-          console.log('Clipboard permission prompt. Not sure what to do here.');
-          break;
-        default:
-          console.log('Clipboard permission not granted.');
-          break;
-      }
-    }
-    catch (error) {
-      console.error('Error pasting from clipboard:', error);
-    }
-  };
-
-
-  const disconnectButton = () => {
-    disconnectClient();
-    setError("");
-  }
 
   return (<main>
     <div className="container">
@@ -362,53 +260,11 @@ export default function Home() {
         }
 
         {(status === 'connected' || chatStarted) &&
-         <div className="messages">
-            <div className="title">Messages:</div>
-            <div className="message-box">
-                <div className="message-panel"></div>
-                <div id="message" className="dark:bg-slate-900">
-                  <pre>
-                    {history.map((msg, index) => (
-                      <div key={index}>
-                        {msg}
-                        {msg !== copiedMessage && (
-                          <button
-                            className="copy-btn"
-                            onClick={() => copyToClipboard(msg)}
-                          >
-                            <BiCopy />
-                          </button>
-                        )}
-                      </div>
-                    ))}
-                  </pre>
-                </div>
-                <input 
-                  type="text"
-                  className="dark:bg-black"
-                  value={message}
-                  onChange={(e) => setMessage(e.target.value)}
-                  placeholder="Enter a message..." 
-                  autoFocus={true}
-                  onKeyDown={(e) => {
-                    if (e.key === "Enter") {
-                      e.preventDefault();
-                      handleSendMessage();
-                    }
-                  }}>
-                </input>
-            </div>
-            <div className="btns">
-                <button className="btn send" onClick={handleSendMessage}>Send</button>
-                <button className="btn clearMsg" onClick={clearMessages}>Clear Messages (Locally)</button>  
-                <button className="btn past-clipbrd" onClick={handlePaste}>Paste clipboard</button>
-                <button className='btn' onClick={() => disconnectButton()}>Disconnect</button>
-                <button className='btn' onClick={handleSendFile}>Send File</button>
-                <input type="file" id="input" onChange={onFileChange} />
-            </div>            
-        </div>}
+        <Chat socket={socket} peerPublicKey={peerPublicKey} 
+        setChatStarted={setChatStarted} disconnectClient={setDisconnect}
+        history={history} setHistory={setHistory}
+        setError={setError}/>
+        }
     </div>
 </main>)
 }
-
-
